LayoutTests/http/tests/paymentrequest/paymentrequest-features.https-expected.txt

Tests for providing `features` as part of `ApplePayRequest`.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ApplePayRequest.features

SETUP: paymentMethod = validPaymentMethod(); delete paymentMethod.data.features
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) did not throw exception.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) did not throw exception.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'lineItemPaymentTiming' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) did not throw exception.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'lineItemPaymentTiming', 'paymentTimingImmediate' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) did not throw exception.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ '' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'invalid' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'null' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 7 ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'lineItemPaymentTiming', '' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'lineItemPaymentTiming', 'invalid' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'lineItemPaymentTiming', 'null' ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = [ 'lineItemPaymentTiming', 7 ]
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = { }
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Type error.

SETUP: paymentMethod = validPaymentMethod(); paymentMethod.data.features = false
PASS new PaymentRequest([paymentMethod], validPaymentDetails()) threw exception TypeError: Value is not a sequence.


PASS successfullyParsed is true

TEST COMPLETE
